pyproject.toml
[project]
name = "aicrawler"
version = "0.2.0"
description = "Weekly AI news briefings - collects, triages, clusters, and narrates AI developments"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "click>=8.0",
    "feedparser>=6.0",
    "flask>=3.0",
    "httpx>=0.27",
    "markdown>=3.5",
    "numpy>=1.24",
    "openai>=1.0",
    "pyyaml>=6.0",
    "python-dotenv>=1.0",
    "scipy>=1.11",
    "sentence-transformers>=2.2.0",
    "trafilatura>=1.6",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "ruff>=0.4",
    "pytest-cov>=4.0",
]

[project.scripts]
aicrawler = "src.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
